first changes for #12
