--- pq/server/__init__.py
@@ -1,11 +1,0 @@
-from .consumer import ConsumerMixin
-from .scheduler import SchedulerMixin
-from .producer import TaskProducer
-
-
-class TaskScheduler(SchedulerMixin, TaskProducer):
-    pass
-
-
-class TaskConsumer(ConsumerMixin, TaskProducer):
-    pass
