+++ pq/server/consumer.py
@@ -1,332 +1,31 @@
-import sys
-import time
-import traceback
-import multiprocessing
-from asyncio import ensure_future, Future, wait_for
-from asyncio import CancelledError, TimeoutError
-from asyncio.subprocess import Process
+import asyncio
 
-from pulsar import CANCELLED_ERRORS
-from pulsar.utils.system import json
-
-from ..tasks.task import TaskError, TaskTimeout
-from ..tasks import models
-from ..tasks import states
-from ..cpubound import StreamProtocol, PROCESS_FILE
-from ..utils.exc import string_exception
-from .pubsub import backoff, RECONNECT_LAG
+from .producer import Producer
 
 
-consumer_event = 'consumer_status'
-
-
-class RemoteStackTrace(TaskError):
-    pass
-
-
-class TooManyTasksForJob(TaskError):
-    status = states.REVOKED
-
-
-class ExecutorMixin:
-    # Mixin for both TaskConsumer and TaskProducer
-    #
-    # The TaskProducer can execute a task inline, while the consumer executes
-    # task from a task queue via the ConsumerMixin
-    #
-    _concurrent_tasks = None
-
-    async def _execute_task(self, task, worker=None):
-        # Function executing a task
-        #
-        # - If the stat_time is greater than task.expiry Revoke the Task
-        # - If the task has a delay not yet reached
-        #   - queue the task at the right time
-        #   - otherwise proceed to next
-        # - Set status to STARTED and consume the task
-        logger = self.logger
-        task_id = task.id
-        time_ended = time.time()
-        job = None
-        JobClass = self.registry.get(task.name)
-
-        try:
-            if not JobClass:
-                raise RuntimeError('%s not in registry' % task.name)
-
-            if task.status > states.STARTED:
-                expiry = task.expiry
-                timeout = None
-
-                if expiry:  # Handle expiry
-                    timeout = expiry - time_ended
-                    if timeout <= 0:
-                        raise TaskTimeout
-
-                if task.delay:  # Task with delay
-                    start_time = task.time_queued + task.delay
-                    gap = start_time - time_ended
-                    if gap > 0:
-                        self._loop.call_later(gap, self._queue_again, task)
-                        if self._concurrent_tasks:
-                            self._concurrent_tasks.pop(task_id, None)
-                        return task
-
-                concurrent = await self.broker.incr(JobClass.name)
-
-                job = JobClass(self, task)
-
-                if job.max_concurrency and concurrent > job.max_concurrency:
-                    raise TooManyTasksForJob('max concurrency %d reached',
-                                             job.max_concurrency)
-
-                kwargs = task.kwargs or {}
-                task.status = states.STARTED
-                task.time_started = time_ended
-                if worker:
-                    task.worker = worker.aid
-                logger.info(task.lazy_info())
-                await self.pubsub.publish('started', task)
-                future = self._consume(job, kwargs)
-                #
-                # record future for cancellation
-                if self._concurrent_tasks:
-                    self._concurrent_tasks[task_id].future = future
-                #
-                # This may block until timeout
-                task.result = await wait_for(future, timeout)
-            else:
-                raise TaskError('Invalid status %s' % task.status_string)
-
-        except (CancelledError, TimeoutError, TaskTimeout):
-            task.result = None
-            task.status = states.REVOKED
-            logger.info(task.lazy_info())
-        except RemoteStackTrace:
-            task.status = states.FAILURE
-            logger.error(task.lazy_info())
-        except TaskError as exc:
-            task.result = string_exception(exc)
-            task.status = exc.status
-            logger.error(task.lazy_info())
-        except Exception as exc:
-            exc_info = sys.exc_info()
-            task.result = string_exception(exc)
-            task.status = states.FAILURE
-            task.stacktrace = traceback.format_tb(exc_info[2])
-            task.exception = traceback.format_exception_only(
-                exc_info[0], exc_info[1])[0]
-            logger.exception(task.lazy_info())
-        else:
-            task.status = states.SUCCESS
-            logger.info(task.lazy_info())
-        #
-        task.time_ended = time.time()
-        if self._concurrent_tasks:
-            self._concurrent_tasks.pop(task_id, None)
-
-        await self.pubsub.publish('done', task)
-
-        if job:
-            await self.broker.decr(job.name)
-            if self._should_retry(job):
-                await self._requeue_task(job)
-
-        return task
-
-    def _consume(self, job, kwargs):
-        concurrency = job.get_concurrency()
-
-        if concurrency == models.THREAD_IO:
-            return job._loop.run_in_executor(None, lambda: job(**kwargs))
-
-        elif concurrency == models.CPUBOUND:
-            return self._consume_in_subprocess(job, kwargs)
-
-        else:
-            return self.green_pool.submit(job, **kwargs)
-
-    async def _consume_in_subprocess(self, job, kwargs):
-        params = dict(self.json_params())
-        loop = job._loop
-
-        transport, protocol = await loop.subprocess_exec(
-            lambda: StreamProtocol(job),
-            sys.executable,
-            PROCESS_FILE,
-            json.dumps(sys.path),
-            json.dumps(params),
-            json.dumps(job.task.tojson()))
-        process = Process(transport, protocol, loop)
-        await process.communicate()
-        if job.task.stacktrace:
-            raise RemoteStackTrace
-        return job.task.result
-
-    def _should_retry(self, job):
-        return (job.task.status != states.SUCCESS and
-                job.task.queue and
-                job.max_retries and
-                job.task.retry < job.max_retries)
-
-    def _requeue_task(self, job):
-        task = job.task
-        meta_params = task.meta.copy()
-        meta_params['retry'] = task.retry + 1
-        return job.queue_task(
-            job.name,
-            callback=False,
-            meta_params=meta_params,
-            queue=task.queue,
-            delay=job.retry_delay,
-            **task.kwargs
-        )
-
-    def _queue_again(self, task):
-        self.broker.queue(task, False)
-
-    def json_params(self):
-        for name, value in self.cfg.items():
-            try:
-                json.dumps(value)
-            except Exception:
-                continue
-            yield name, value
-
-
-class ConsumerMixin:
-    """A mixin for consuming tasks from a distributed task queue.
-    """
-    @classmethod
-    def __new__(cls, *args, **kwargs):
-        o = super().__new__(cls)
-        o._processed = 0
-        o._next_time = 1
-        o._closing_waiter = None
-        o._concurrent_tasks = {}
-        return o
+class Consumer(Producer):
 
     def __repr__(self):
-        return 'task consumer <%s>' % self.broker
-
-    def queues(self):
-        '''List of task queues consumed by this task consumer
-        '''
-        return self.manager.queues()
+        return 'consumer <%s>' % self.broker
 
     @property
-    def num_concurrent_tasks(self):
-        '''The number of concurrent_tasks
-        '''
-        return len(self._concurrent_tasks)
+    def is_consumer(self):
+        return True
 
-    def info(self):
-        return {'node': self.node_name,
-                'concurrent': list(self._concurrent_tasks),
-                'processed': self._processed,
-                'queues': self.queues()}
+    def tick(self):
+        for consumer in self.consumers:
+            consumer.tick()
 
-    async def start(self, worker):
-        '''Starts consuming tasks
-        '''
+    async def start(self, worker=None):
         await self.pubsub.start()
-        self._poll_tasks(worker)
-        self.logger.warning('%s started polling tasks', self)
+        for consumer in self.consumers:
+            consumer.start(worker)
         return self
 
-    def closing(self):
-        return self._closing_waiter is not None
-
-    def close(self, msg=None):
+    async def close(self, msg=None):
         if not self.closing():
-            self._closing_waiter = Future(loop=self._loop)
-            self.logger.warning(msg)
-        return self._closing_waiter
-
-    # #######################################################################
-    # #    PRIVATE METHODS
-    # #######################################################################
-    def _poll_tasks(self, worker, next_time=None):
-        if self.closing() and not self._concurrent_tasks:
-            self._do_close(worker)
-        elif worker.is_running() and not next_time:
-            ensure_future(self._may_poll_task(worker), loop=worker._loop)
-        else:
-            next_time = next_time or 0
-            worker._loop.call_later(next_time, self._poll_tasks, worker)
-
-    async def _may_poll_task(self, worker):
-        # Called in the ``worker`` event loop.
-        #
-        # It pools a new task if possible, and add it to the queue of
-        # tasks consumed by the ``worker`` CPU-bound thread.'''
-        next_time = None
-        if worker.is_running():
-            cpu_count = multiprocessing.cpu_count()
-            concurrent_tasks = cpu_count*self.cfg.concurrent_tasks
-
-            # executor = worker.executor()
-            if self.num_concurrent_tasks < concurrent_tasks:
-                max_tasks = self.cfg.max_requests
-                if max_tasks and self._processed >= max_tasks:
-                    self.close(
-                        'Processed %s tasks. Stop polling tasks.'
-                        % self._processed
-                    )
-
-                if not self.closing():
-                    try:
-                        task = await self.broker.get_task(*self.queues())
-                    except ConnectionRefusedError:
-                        if self.broker.connection_error:
-                            self._next_time = backoff(self._next_time)
-                        else:
-                            self._next_time = RECONNECT_LAG
-                            self.broker.connection_error = True
-                        next_time = self._next_time
-                        task = None
-                        if worker.is_running():
-                            self.logger.critical(
-                                '%s cannot pool tasks - '
-                                'connection error - try again in %s seconds',
-                                self.broker,
-                                next_time
-                            )
-                    except CANCELLED_ERRORS:
-                        self.logger.debug('stopped polling tasks')
-                        raise
-                    else:
-                        self.broker.connection_ok()
-                    if task:  # Got a new task
-                        self._processed += 1
-                        self._concurrent_tasks[task.id] = TaskExecutor(task)
-                        ensure_future(self._execute_task(task, worker))
-                    await self._broadcast(worker)
-            else:
-                self.logger.debug('%s concurrent tasks. Cannot poll.',
-                                  concurrent_tasks)
-                self._next_time = 1
-                next_time = self._next_time
-                await self._broadcast(worker)
-
-        self._poll_tasks(worker, next_time)
-
-    def _broadcast(self, worker):
-        info = self.info()
-        info['worker'] = worker.aid
-        info['time'] = time.time()
-        return self.pubsub.publish(consumer_event, info)
-
-    def _do_close(self, worker):
-        self.logger.warning('Closing %s', self)
-        self.manager.close()
-        self._closing_waiter.set_result(True)
-        if not worker.is_monitor():
-            self._loop.call_later(1, self._loop.stop)
-
-
-class TaskExecutor:
-
-    def __init__(self, task):
-        self.task = task
-        self.future = None
+            closing = []
+            for consumer in self.consumers:
+                closing.append(consumer.close())
+            await asyncio.gather(*closing)
+            self.manager.close()
